* re.c (KR_REHASH): wrong hash value on sizeof(long) > sizeof(int).
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@5693 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ re.c
@@ -105,7 +105,7 @@ rb_memsearch(x0, m, y0, n)
     int d;
     unsigned long hx, hy;
 
-#define KR_REHASH(a, b, h) (((h) << 1) - ((a)<<d) + (b))
+#define KR_REHASH(a, b, h) (((h) << 1) - ((long)(a)<<d) + (b))
 
     if (m > n) return -1;
     s = y; e = s + n - m;
